+++ src/klib/khash.h
