+++ src/klib/kvec.h
