+++ src/nvim/os/pty_conpty_win.h
@@ -1,7 +1,7 @@
 #ifndef NVIM_OS_PTY_CONPTY_WIN_H
 #define NVIM_OS_PTY_CONPTY_WIN_H
 
-#include "nvim/lib/kvec.h"
+#include "klib/kvec.h"
 #include "nvim/os/input.h"
 
 #ifndef HPCON
