+++ src/nvim/strings.h
@@ -5,8 +5,8 @@
 #include <stdbool.h>
 #include <string.h>
 
+#include "klib/kvec.h"
 #include "nvim/eval/typval.h"
-#include "nvim/lib/kvec.h"
 #include "nvim/types.h"
 
 /// Append string to string and return pointer to the next byte
